Split ingestion tests into dedicated file and add YFinanceMarketData tests
Moves NewsAPIProvider and RedditProvider tests from test_sentiment_logic
into a new test_ingestion.py alongside new YFinanceMarketData tests.
test_sentiment_logic.py now covers only FinBERTModel and safe().
